chore: move IssueRemover into class file and refactor script entry
